Show context length for running models
When the model is running, next to the name of the model on line 1, show with a tertiary color the context length at which the model is running. The idea is that when installed models are idle, we show model properties, but when they are running, we can start showing properties of this particular run.